Move APISpec
