+++ gateway/api_definition.go
@@ -10,11 +10,9 @@ import (
 	"io"
 	"io/ioutil"
 	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
 	"strings"
-	"sync"
 	"sync/atomic"
 	texttemplate "text/template"
 	"time"
@@ -22,9 +20,6 @@ import (
 	"github.com/TykTechnologies/tyk/ee/middleware/streams"
 	"github.com/TykTechnologies/tyk/storage/kv"
 
-	"github.com/getkin/kin-openapi/routers"
-
-	"github.com/TykTechnologies/tyk/internal/graphengine"
 	"github.com/TykTechnologies/tyk/internal/httputil"
 
 	"github.com/getkin/kin-openapi/routers/gorillamux"
@@ -42,8 +37,6 @@ import (
 	circuit "github.com/TykTechnologies/circuitbreaker"
 
 	"github.com/TykTechnologies/gojsonschema"
-
-	"github.com/TykTechnologies/tyk-pump/analytics"
 
 	"github.com/TykTechnologies/tyk/apidef"
 	"github.com/TykTechnologies/tyk/config"
@@ -151,53 +144,6 @@ type ExtendedCircuitBreakerMeta struct {
 
 type OAuthManagerInterface interface {
 	Storage() ExtendedOsinStorageInterface
-}
-
-// APISpec represents a path specification for an API, to avoid enumerating multiple nested lists, a single
-// flattened URL list is checked for matching paths and then it's status evaluated if found.
-type APISpec struct {
-	*apidef.APIDefinition
-	OAS oas.OAS
-
-	sync.RWMutex
-
-	Checksum         string
-	RxPaths          map[string][]URLSpec
-	WhiteListEnabled map[string]bool
-	target           *url.URL
-	AuthManager      SessionHandler
-	OAuthManager     OAuthManagerInterface
-
-	OrgSessionManager        SessionHandler
-	EventPaths               map[apidef.TykEvent][]config.TykEventHandler
-	Health                   HealthChecker
-	JSVM                     JSVM
-	ResponseChain            []TykResponseHandler
-	RoundRobin               RoundRobin
-	URLRewriteEnabled        bool
-	CircuitBreakerEnabled    bool
-	EnforcedTimeoutEnabled   bool
-	LastGoodHostList         *apidef.HostList
-	HasRun                   bool
-	ServiceRefreshInProgress bool
-	HTTPTransport            *TykRoundTripper
-	HTTPTransportCreated     time.Time
-	WSTransport              http.RoundTripper
-	WSTransportCreated       time.Time
-	GlobalConfig             config.Config
-	OrgHasNoSession          bool
-	AnalyticsPluginConfig    *GoAnalyticsPlugin
-
-	middlewareChain *ChainObject
-	unloadHooks     []func()
-
-	network analytics.NetworkStats
-
-	GraphEngine graphengine.Engine
-
-	HasMock            bool
-	HasValidateRequest bool
-	OASRouter          routers.Router
 }
 
 // GetSessionLifetimeRespectsKeyExpiration returns a boolean to tell whether session lifetime should respect to key expiration or not.
