+++ gateway/model_apispec.go
@@ -2,8 +2,65 @@ package gateway
 
 import (
 	"net/http"
+	"net/url"
 	"strings"
+	"sync"
+	"time"
+
+	"github.com/TykTechnologies/tyk-pump/analytics"
+	"github.com/TykTechnologies/tyk/apidef"
+	"github.com/TykTechnologies/tyk/apidef/oas"
+	"github.com/TykTechnologies/tyk/config"
+	"github.com/TykTechnologies/tyk/internal/graphengine"
+	"github.com/getkin/kin-openapi/routers"
 )
+
+// APISpec represents a path specification for an API, to avoid enumerating multiple nested lists, a single
+// flattened URL list is checked for matching paths and then it's status evaluated if found.
+type APISpec struct {
+	*apidef.APIDefinition
+	OAS oas.OAS
+
+	sync.RWMutex
+
+	Checksum         string
+	RxPaths          map[string][]URLSpec
+	WhiteListEnabled map[string]bool
+	target           *url.URL
+	AuthManager      SessionHandler
+	OAuthManager     OAuthManagerInterface
+
+	OrgSessionManager        SessionHandler
+	EventPaths               map[apidef.TykEvent][]config.TykEventHandler
+	Health                   HealthChecker
+	JSVM                     JSVM
+	ResponseChain            []TykResponseHandler
+	RoundRobin               RoundRobin
+	URLRewriteEnabled        bool
+	CircuitBreakerEnabled    bool
+	EnforcedTimeoutEnabled   bool
+	LastGoodHostList         *apidef.HostList
+	HasRun                   bool
+	ServiceRefreshInProgress bool
+	HTTPTransport            *TykRoundTripper
+	HTTPTransportCreated     time.Time
+	WSTransport              http.RoundTripper
+	WSTransportCreated       time.Time
+	GlobalConfig             config.Config
+	OrgHasNoSession          bool
+	AnalyticsPluginConfig    *GoAnalyticsPlugin
+
+	middlewareChain *ChainObject
+	unloadHooks     []func()
+
+	network analytics.NetworkStats
+
+	GraphEngine graphengine.Engine
+
+	HasMock            bool
+	HasValidateRequest bool
+	OASRouter          routers.Router
+}
 
 // CheckSpecMatchesStatus checks if a URL spec has a specific status.
 // Deprecated: The function doesn't follow go return conventions (T, ok); use FindSpecMatchesStatus;
